Update d3js-converter.py
Works!,now in this update, the source generates the json file correctly. If you look, now are not created the three files csv for saving the data of the columns, as there were replaced with a for loop that amending a index of the variable that contains the data of the celds (colb[celd], cold[celd], colf[celd]).
The code is smaller than the previous, and the performance is better.